refactor: ClubApi 패키지 변경
